Safari-611.1.10/LayoutTests/http/tests/contentextensions/hide-on-csp-report-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost/foo.png because it does not appear in the img-src directive of the Content Security Policy.
This test creates a CSP violation report, but the report URL matches a 'css-display-none' rule.
This text should remain visible.



--------
Frame: 'result_frame'
--------
Ping received.
HTTP_HOST: localhost:8000
REQUEST_URI: /contentextensions/resources/save-ping.php?test=contentextensions-hide-on-csp-report
No cookies in ping.
